requirements.txt
tensorflow-macos
tensorflow-metal
tensorflow-hub
torch
torchvision
transformers
scikit-learn
matplotlib
pillow
pandas